@@ -11,9 +11,9 @@
     "bootstrap-icons": "^1.7.2",
     "disable-scroll": "^0.6.0",
     "prop-types": "^15.8.0",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.5.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-ga": "^3.3.1",
     "react-ga4": "^1.4.1",
     "react-router-dom": "^5.3.3",
